refactor(file-router): get route title from component if there is no config title.
